pyproject.toml
[tool.poetry]
name = "youtube-sort"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
oauth2client = "^4.1.3"
google-auth = "^2.25.2"
google-auth-oauthlib = "^1.2.0"
google-api-python-client = "^2.111.0"
ruff = "^0.1.14"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
